Add scheduler to save menu to cache every 5 minutes
